move intersection logic to ChunkedEntities
