+++ src/3d/qgsvectorlayerchunkloader_p.h
@@ -120,6 +120,8 @@ class QgsVectorLayerChunkedEntity : public QgsChunkedEntity
     //! Constructs the entity. The argument maxLevel determines how deep the tree of tiles will be
     explicit QgsVectorLayerChunkedEntity( QgsVectorLayer *vl, double zMin, double zMax, const QgsVectorLayer3DTilingSettings &tilingSettings, QgsAbstract3DSymbol *symbol, const Qgs3DMapSettings &map );
 
+    QVector<RayHit> intersectEntity( const QgsRay3D &ray, const RayCastContext &context ) const override;
+
     ~QgsVectorLayerChunkedEntity();
   private slots:
     void onTerrainElevationOffsetChanged( float newOffset );
