+++ src/3d/terrain/qgsterrainentity_p.h
@@ -51,6 +51,8 @@ class QgsCoordinateTransform;
 class QgsMapLayer;
 class QgsTerrainGenerator;
 class TerrainMapUpdateJobFactory;
+class QgsRay3D;
+class RayHit;
 
 /**
  * \ingroup 3d
@@ -81,6 +83,8 @@ class QgsTerrainEntity : public QgsChunkedEntity
 
     //! Tests whether the ray intersects the terrain - if it does, it sets the intersection point (in world coordinates)
     bool rayIntersection( const QgsRayCastingUtils::Ray3D &ray, QVector3D &intersectionPoint );
+
+    QVector<RayHit> intersectEntity( const QgsRay3D &ray, const RayCastContext &context ) const override;
 
   private slots:
     void onShowBoundingBoxesChanged();
